# modules/guestbook/guestbook-impl/rest-openapi.yaml
components:
    schemas:
        Guest:
            properties:
                guestId:
                    format: int64
                    type: integer
                name:
                    type: string
            type: object
info:
    description:
        "Guestbook REST API. A Java client JAR is available for use with the group ID 'com.liferay', artifact ID
        'com.liferay.guestbook.client', and version '1.0.0'."
    license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
    title: "Guestbook"
    version: v1.0
openapi: 3.0.1
paths:
    /guests:
        get:
            description:
                Retrieves the guests
            operationId: getGuestsPage
            responses:
                default:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/Guest"
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: "#/components/schemas/Guest"
                                type: array
                    description:
                        default response
            tags: ["Guest"]
        post:
            description:
                Adds a new guest
            operationId: postGuest
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/Guest"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/Guest"
            responses:
                default:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Guest"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Guest"
                    description:
                        default response
            tags: ["Guest"]
    /guests/{guestId}:
        get:
            description:
                Retrieves the guest with given identifier
            operationId: getGuest
            parameters:
                - in: path
                  name: guestId
                  required: true
                  schema:
                      type: string
            responses:
                default:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Guest"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Guest"
                    description:
                        default response
            tags: ["Guest"]